Remove deprecated dependency nom v1.2.4
We don't use this dependency, it was pulled by mistake through
an unmaintained package steamy-vdf.  Package steamlocate just merged a
fix removing dependency on that unmaintained library and is preparing
for v2.0 release - fortunately for us, we can just drop it as-is,
without any further changes.

This commit fixes warning that keeps popping up on every build with
rustc 1.68.0:

> warning: the following packages contain code that will be rejected
> by a future version of Rust: nom v1.2.4